+++ integration/test_collection_rerank.py
@@ -4,16 +4,14 @@ import pytest
 
 import weaviate.classes as wvc
 from weaviate.util import _ServerVersion
-
 from .conftest import CollectionFactory
 
 
 def test_query_using_rerank_with_old_server(collection_factory: CollectionFactory) -> None:
     collection = collection_factory(
-        reranker_config=wvc.config.Configure.Reranker.transformers(),
+        reranker_config=wvc.config.Configure.Reranker.custom("reranker-dummy"),
         vectorizer_config=wvc.config.Configure.Vectorizer.none(),
         properties=[wvc.config.Property(name="text", data_type=wvc.config.DataType.TEXT)],
-        ports=(8079, 50050),
     )
     if collection._connection._weaviate_version >= _ServerVersion(1, 23, 1):
         pytest.skip("Reranking works with 1.23.1 or higher so no need to test this")
@@ -36,10 +34,9 @@ def test_queries_with_rerank(collection_factory: CollectionFactory) -> None:
 
     collection = collection_factory(
         name="Test_test_queries_with_rerank",
-        reranker_config=wvc.config.Configure.Reranker.transformers(),
+        reranker_config=wvc.config.Configure.Reranker.custom("reranker-dummy"),
         vectorizer_config=wvc.config.Configure.Vectorizer.text2vec_openai(),
         properties=[wvc.config.Property(name="text", data_type=wvc.config.DataType.TEXT)],
-        ports=(8079, 50050),
         headers={"X-OpenAI-Api-Key": api_key},
     )
     if collection._connection._weaviate_version < _ServerVersion(1, 23, 1):
@@ -92,12 +89,11 @@ def test_queries_with_rerank_and_group_by(collection_factory: CollectionFactory)
 
     collection = collection_factory(
         name="Test_test_queries_with_rerank_and_group_by",
-        reranker_config=wvc.config.Configure.Reranker.transformers(),
+        reranker_config=wvc.config.Configure.Reranker.custom("reranker-dummy"),
         vectorizer_config=wvc.config.Configure.Vectorizer.text2vec_openai(
             vectorize_collection_name=False
         ),
         properties=[wvc.config.Property(name="text", data_type=wvc.config.DataType.TEXT)],
-        ports=(8079, 50050),
         headers={"X-OpenAI-Api-Key": api_key},
     )
     if collection._connection._weaviate_version < _ServerVersion(1, 23, 1):