+++ test/collection/test_config.py
@@ -83,6 +83,40 @@ TEST_CONFIG_WITH_VECTORIZER_PARAMETERS = [
         },
     ),
     (
+        Configure.Vectorizer.multi2vec_cohere(
+            model="embed-multilingual-v2.0",
+            truncate="NONE",
+            vectorize_collection_name=False,
+            base_url="https://api.cohere.ai",
+        ),
+        {
+            "multi2vec-cohere": {
+                "model": "embed-multilingual-v2.0",
+                "truncate": "NONE",
+                "vectorizeClassName": False,
+                "baseURL": "https://api.cohere.ai/",
+            }
+        },
+    ),
+    (
+        Configure.Vectorizer.multi2vec_voyageai(
+            model="voyage-multimodal-3",
+            truncation=False,
+            output_encoding="base64",
+            vectorize_collection_name=False,
+            base_url="https://api.voyageai.com",
+        ),
+        {
+            "multi2vec-voyageai": {
+                "model": "voyage-multimodal-3",
+                "truncation": False,
+                "output_encoding": "base64",
+                "vectorizeClassName": False,
+                "baseURL": "https://api.voyageai.com/",
+            }
+        },
+    ),
+    (
         Configure.Vectorizer.text2vec_gpt4all(),
         {
             "text2vec-gpt4all": {
@@ -340,6 +374,34 @@ TEST_CONFIG_WITH_VECTORIZER_PARAMETERS = [
         },
     ),
     (
+        Configure.Vectorizer.text2vec_jinaai(
+            model="jina-embeddings-v3",
+            vectorize_collection_name=False,
+            dimensions=512,
+        ),
+        {
+            "text2vec-jinaai": {
+                "model": "jina-embeddings-v3",
+                "vectorizeClassName": False,
+                "dimensions": 512,
+            }
+        },
+    ),
+    (
+        Configure.Vectorizer.multi2vec_jinaai(
+            model="jina-clip-v2",
+            dimensions=512,
+            vectorize_collection_name=False,
+        ),
+        {
+            "multi2vec-jinaai": {
+                "model": "jina-clip-v2",
+                "dimensions": 512,
+                "vectorizeClassName": False,
+            }
+        },
+    ),
+    (
         Configure.Vectorizer.text2vec_voyageai(
             vectorize_collection_name=False,
             model="voyage-large-2",
@@ -352,6 +414,22 @@ TEST_CONFIG_WITH_VECTORIZER_PARAMETERS = [
                 "model": "voyage-large-2",
                 "baseURL": "https://voyage.made-up.com",
                 "truncate": False,
+            }
+        },
+    ),
+    (
+        Configure.Vectorizer.text2vec_weaviate(
+            vectorize_collection_name=False,
+            model="Snowflake/snowflake-arctic-embed-m-v1.5",
+            base_url="https://api.embedding.weaviate.io",
+            dimensions=768,
+        ),
+        {
+            "text2vec-weaviate": {
+                "vectorizeClassName": False,
+                "model": "Snowflake/snowflake-arctic-embed-m-v1.5",
+                "baseURL": "https://api.embedding.weaviate.io",
+                "dimensions": 768,
             }
         },
     ),
@@ -1220,6 +1298,53 @@ TEST_CONFIG_WITH_NAMED_VECTORIZER_PARAMETERS = [
         },
     ),
     (
+        [Configure.NamedVectors.multi2vec_cohere(name="test", text_fields=["prop"])],
+        {
+            "test": {
+                "vectorizer": {
+                    "multi2vec-cohere": {
+                        "vectorizeClassName": True,
+                        "textFields": ["prop"],
+                    }
+                },
+                "vectorIndexType": "hnsw",
+            }
+        },
+    ),
+    (
+        [Configure.NamedVectors.multi2vec_voyageai(name="test", text_fields=["prop"])],
+        {
+            "test": {
+                "vectorizer": {
+                    "multi2vec-voyageai": {
+                        "vectorizeClassName": True,
+                        "textFields": ["prop"],
+                    }
+                },
+                "vectorIndexType": "hnsw",
+            }
+        },
+    ),
+    (
+        [
+            Configure.NamedVectors.multi2vec_jinaai(
+                name="test", dimensions=256, text_fields=["prop"]
+            )
+        ],
+        {
+            "test": {
+                "vectorizer": {
+                    "multi2vec-jinaai": {
+                        "dimensions": 256,
+                        "vectorizeClassName": True,
+                        "textFields": ["prop"],
+                    }
+                },
+                "vectorIndexType": "hnsw",
+            }
+        },
+    ),
+    (
         [Configure.NamedVectors.text2vec_gpt4all(name="test", source_properties=["prop"])],
         {
             "test": {
@@ -1412,6 +1537,29 @@ TEST_CONFIG_WITH_NAMED_VECTORIZER_PARAMETERS = [
                         "properties": ["prop"],
                         "vectorizeClassName": True,
                         "truncate": True,
+                    }
+                },
+                "vectorIndexType": "hnsw",
+            }
+        },
+    ),
+    (
+        [
+            Configure.NamedVectors.text2vec_weaviate(
+                name="test",
+                source_properties=["prop"],
+                base_url="https://api.embedding.weaviate.io",
+                dimensions=768,
+            )
+        ],
+        {
+            "test": {
+                "vectorizer": {
+                    "text2vec-weaviate": {
+                        "properties": ["prop"],
+                        "vectorizeClassName": True,
+                        "baseURL": "https://api.embedding.weaviate.io",
+                        "dimensions": 768,
                     }
                 },
                 "vectorIndexType": "hnsw",