+++ weaviate/classification/__init__.py
@@ -1,7 +1,0 @@
-"""
-Module for classifying objects within Weaviate.
-"""
-
-__all__ = ["Classification", "ConfigBuilder"]
-
-from .classification import Classification, ConfigBuilder