+++ weaviate/collections/batch/grpc_batch_objects.py
@@ -4,8 +4,8 @@ import time
 import uuid as uuid_package
 from typing import Any, Dict, List, Optional, Union, cast
 
-from grpc.aio import AioRpcError  # type: ignore
 from google.protobuf.struct_pb2 import Struct
+from grpc.aio import AioRpcError  # type: ignore
 
 from weaviate.collections.classes.batch import (
     ErrorObject,
@@ -14,15 +14,16 @@ from weaviate.collections.classes.batch import (
     BatchObjectReturn,
 )
 from weaviate.collections.classes.config import ConsistencyLevel
-from weaviate.collections.classes.types import GeoCoordinate, PhoneNumber
 from weaviate.collections.classes.internal import ReferenceToMulti, ReferenceInputs
-from weaviate.collections.grpc.shared import _BaseGRPC
+from weaviate.collections.classes.types import GeoCoordinate, PhoneNumber
+from weaviate.collections.grpc.shared import _BaseGRPC, PERMISSION_DENIED
 from weaviate.connect import ConnectionV4
 from weaviate.exceptions import (
     WeaviateBatchError,
     WeaviateInsertInvalidPropertyError,
     WeaviateInsertManyAllFailedError,
     WeaviateInvalidInputError,
+    InsufficientPermissionsError,
 )
 from weaviate.proto.v1 import batch_pb2, base_pb2
 from weaviate.util import _datetime_to_string, _get_vector_v4
@@ -138,7 +139,6 @@ class _BatchGRPC(_BaseGRPC):
     async def __send_batch(
         self, batch: List[batch_pb2.BatchObject], timeout: Union[int, float]
     ) -> Dict[int, str]:
-        metadata = self._get_metadata()
         try:
             assert self._connection.grpc_stub is not None
             res = await self._connection.grpc_stub.BatchObjects(
@@ -146,7 +146,7 @@ class _BatchGRPC(_BaseGRPC):
                     objects=batch,
                     consistency_level=self._consistency_level,
                 ),
-                metadata=metadata,
+                metadata=self._connection.grpc_headers(),
                 timeout=timeout,
             )
             res = cast(batch_pb2.BatchObjectsReply, res)
@@ -156,6 +156,8 @@ class _BatchGRPC(_BaseGRPC):
                 objects[result.index] = result.error
             return objects
         except AioRpcError as e:
+            if e.code().name == PERMISSION_DENIED:
+                raise InsufficientPermissionsError(e)
             raise WeaviateBatchError(str(e)) from e
 
     def __translate_properties_from_python_to_grpc(