+++ weaviate/collections/classes/config.py
@@ -137,6 +137,8 @@ class Tokenization(str, Enum):
             Tokenize using GSE (for Chinese and Japanese).
         `TRIGRAM`
             Tokenize into trigrams.
+        `KAGOME_JA`
+            Tokenize using the 'Kagome' tokenizer (for Japanese).
         `KAGOME_KR`
             Tokenize using the 'Kagome' tokenizer and a Korean MeCab dictionary (for Korean).
     """
@@ -147,6 +149,7 @@ class Tokenization(str, Enum):
     FIELD = "field"
     GSE = "gse"
     TRIGRAM = "trigram"
+    KAGOME_JA = "kagome_ja"
     KAGOME_KR = "kagome_kr"
 
 
@@ -243,6 +246,7 @@ class ReplicationDeletionStrategy(str, Enum):
 
     DELETE_ON_CONFLICT = "DeleteOnConflict"
     NO_AUTOMATED_RESOLUTION = "NoAutomatedResolution"
+    TIME_BASED_RESOLUTION = "TimeBasedResolution"
 
 
 class PQEncoderType(str, Enum):